Change detection logic
Code now doesn't run every frame. Only runs if the ray scan detects something valid. Most of the logic here was borrowed from the world tool tips mod so shoutout to that. This also fixes an issue with the world tool tips mod that seemed to cause a race condition where doors would have the wrong properties such as having the names of a different building or the schedule of a different building. Code not running every frame means it's possible to manually set the mode after looking at an object if it isn't correct.

Change to steal mode if the player is looking at a valid object such as an NPC or door while sneaking.

Handle a lot of other mods for most of the objects in Daggerfall. I was able to complete a playthrough without setting it manually at any point I think.

Add code to use levitation movement logic while swimming. Just nicer than using the page up/down keys for ascending and descending in water in my opinion. You should still sink like a rock if you have too many items.